Code/Backend/linkedin_scraper/requirements.txt
selenium
requests
lxml
python-dotenv
pandas